Reworked of how actions are enables and disabled in GUI
This commit changes the way gui actions are enabled/disables 
based on a new variable GUIMode. Currently there are noNet mode (when
no net is opened) Draw (when we are in editr/draw mode) and simulation 
(for simulation) 

The old function setAnimationMode() has been marked as depricated, and
is replaced with setGUIMode()

All action not aviable when no net is opened are now disabled, and 
reenabled/disabled based on the current GUIMode.