Added functions for control visibility of element in left menu
I have added some code that makes it possible to hide elements from the 
left menu, eg. the query or constants. Right now the change is globle
but we can make it dependent on the tab. 
Also the elements in the view menu need to be modified to show if you 
enable or disable by clicking it.