Changed the way an annotation is copied, fixes bug:873427
Before when copying an anotation, the placement coordinates of the screen where used, and passed to a function that calculated the unzoomed coordinates. This was wrong and made the new annotation being placed by some offset.

Insted the originalX and originalY of the original annotiation is used as parameters for the new annotation.