Changed Annotation offset on create, fixes bug 873427
After a new annotation was created is was shifted 20px up and 20px left.
When making a copy, the new annotation get shifted again, and thus "moves".

However there is  no good reason to as we see it to make this shifting of the annotation, so this has been removed.

Pair programmed with ~yrke.