Added, zoom "loading/update" functionality when setting the model of the DrawinSurface(DS), fixes bug 873427.
When selecting a PN-component, the corresponding net is set as the model for the DS. But setting the model for the DS, did not update the zoom.

This problem was introduced when adding the functionallity to copy components. Before models where only loaded from file, or as a new component, both always with zoom level 100%.

But when copying a loaded component, it might not be at zoom level 100%, the new compenent had the right zoom level, but it was not loaded correctly.

The graphical element where copied with a zoomed position, this did not work, as the graphical size of the elements where not "zoomed", and the position did not correspond to the zoom level.

Now the graphical elements places, transitions and notes are copied with the original position, i.e. at zoom level 100%, and then they are updated when drawn on the DS.