Removed code to create DrawingSurface if not found
The code should never run, as there is always a DrawingSurface, if not
it seems wrong just to remove it. 
Code left in place to for history, in case we someday experience problems with this
If no problems have been observed for some time, please remove the code and comment.