Removed unneeded static ignoreSelection from PetriNetObject
used to not draw components as selected when in animation or export mode.
However other changes makes sure that you cant select in animation mode, 
so this is not needed any more.

In export mode, the changes is only relevant to png and ps export,
for now i'm changing it to include seleciton in the exports, as I think 
that can be a nice feature. But can be refactored to get original 
behaviour