AnimationController now saves state of tooltip, insted of reading global state
AnimationController enabled tooltop for TimeDelay button to show possible errors, not ideal, but works.
In theory this change could be funky if you disable tooltips while hovering the button chaning it 
back to a wrong state, however testing this it does not seem to be a problem. Also enabeling tooltip
seems a little funky, as quick mouse moves can make the tooltip appere on other elements.
If its a problem we should use something else that tooltip to communicate to the user.