Fix for bug lp:1987383, updating collection while iterating over it
this code seem way out of place and wrong, the way the the update is done is wired.
I guess the  update is done due to a de-sync between the tapnNetwork and gui model, but fixing it this way seems
crazy. Better would be to make sure the correct gui model is passed.
The loop lookes even weirder now, but for the bugfix I just rewrote the loop to keep the same behaviour as before